backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
async-lru==2.0.3
binaryornot==0.4.4
bitarray==2.8.3
bitstring==4.1.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.3
cookiecutter==2.6.0
cramjam==2.8.3
cuser==0.0.4
debugpy==1.6.7
Django==4.1.13
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.4
django-cors-headers==4.3.1
django-cuser==2017.3.16
django-debug-toolbar==4.3.0
django-filter==24.2
django-jazzmin==2.6.0
django-jet==1.0.8
django-modeltranslation==0.18.11
django-mongoengine==0.5.6
django-parler==2.3
django-rosetta==0.10.0
django-tailwind==3.8.0
django-tinymce==3.7.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
esptool==4.6.2
fastjsonschema==2.17.1
fqdn==1.5.1
idna==3.6
ipykernel==6.24.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
Jinja2==3.1.3
json5==0.9.14
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.1
modeltranslation==0.25
mongoengine==0.28.2
mysqlclient==2.2.1
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
notebook==7.0.0
notebook_shim==0.2.3
overrides==7.3.1
pandocfilters==1.5.0
pillow==10.2.0
pipkin==2.0b2
polib==1.2.0
prometheus-client==0.17.1
Pygments==2.17.2
pykerberos==1.2.4
pymongo==3.12.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-slugify==8.0.4
python-snappy==0.7.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rcssmin==1.1.1
reedsolo==1.7.0
referencing==0.30.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rjsmin==1.2.1
rpds-py==0.9.2
six==1.16.0
sqlparse==0.2.4
terminado==0.17.1
text-unidecode==1.3
traitlets==5.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
uri-template==1.3.0
urllib3==2.2.1
whitenoise==6.6.0
widgetsnbextension==4.0.8
